telegram_bot/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
telethon==1.34.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.3.0
